src/gui/requirements.txt
pyzmq==22.0.2
pyqt>=5.0.0
